Move create-event into tabs so nav bar persists; wire Who's Posting to real profile/org data
