--- app/(create-event)/_layout.tsx
@@ -1,10 +1,0 @@
-import { Stack } from 'expo-router';
-import { CreateEventProvider } from '@/app/context/CreateEventContext';
-
-export default function CreateEventLayout() {
-  return (
-    <CreateEventProvider>
-      <Stack screenOptions={{ headerShown: false }} />
-    </CreateEventProvider>
-  );
-}
