--- app/(tabs)/saved.tsx
@@ -1,23 +1,0 @@
-import { StyleSheet, Text, View } from 'react-native';
-
-export default function SavedScreen() {
-  return (
-    <View style={styles.container}>
-      <Text style={styles.title}>Saved Page</Text>
-      <Text>This is the placeholder saved screen.</Text>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-  },
-  title: {
-    fontSize: 24,
-    fontWeight: '600',
-    marginBottom: 8,
-  },
-});
