+++ app/context/CreateEventContext.tsx
@@ -61,10 +61,31 @@ export interface CreateEventData {
   imageMimeType: string | null;
 }
 
+// The six steps of the flow, in order. The create tab renders the component
+// for the current step; advancing/going back moves through this list rather
+// than pushing routes, so the tab bar stays mounted the whole time.
+export const CREATE_EVENT_STEPS = [
+  'whosPosting',
+  'discoveryBucket',
+  'interestTags',
+  'eventDetails',
+  'whenIsIt',
+  'optionalExtras',
+] as const;
+
+export type CreateEventStep = (typeof CREATE_EVENT_STEPS)[number];
+
+export const CREATE_EVENT_STEP_COUNT = CREATE_EVENT_STEPS.length;
+
 interface CreateEventContextType {
   data: CreateEventData;
   update: (partial: Partial<CreateEventData>) => void;
   reset: () => void;
+  /** Index into CREATE_EVENT_STEPS of the step currently shown. */
+  stepIndex: number;
+  step: CreateEventStep;
+  goNext: () => void;
+  goBack: () => void;
 }
 
 const DEFAULT_DATA: CreateEventData = {
@@ -88,19 +109,41 @@ const CreateEventContext = createContext<CreateEventContextType>({
   data: DEFAULT_DATA,
   update: () => {},
   reset: () => {},
+  stepIndex: 0,
+  step: CREATE_EVENT_STEPS[0],
+  goNext: () => {},
+  goBack: () => {},
 });
 
 export function CreateEventProvider({ children }: { children: React.ReactNode }) {
   const [data, setData] = useState<CreateEventData>(DEFAULT_DATA);
+  const [stepIndex, setStepIndex] = useState(0);
 
   const update = (partial: Partial<CreateEventData>) => {
     setData((prev) => ({ ...prev, ...partial }));
   };
 
-  const reset = () => setData(DEFAULT_DATA);
+  // Reset clears both the draft and the step, so the next visit starts fresh.
+  const reset = () => {
+    setData(DEFAULT_DATA);
+    setStepIndex(0);
+  };
+
+  const goNext = () => setStepIndex((i) => Math.min(i + 1, CREATE_EVENT_STEPS.length - 1));
+  const goBack = () => setStepIndex((i) => Math.max(i - 1, 0));
 
   return (
-    <CreateEventContext.Provider value={{ data, update, reset }}>
+    <CreateEventContext.Provider
+      value={{
+        data,
+        update,
+        reset,
+        stepIndex,
+        step: CREATE_EVENT_STEPS[stepIndex],
+        goNext,
+        goBack,
+      }}
+    >
       {children}
     </CreateEventContext.Provider>
   );
